Access ciontrol User cannot acccess articles without Log in Access Control for the Edit/Delete buttons
 Changes to be committed:
	modified:   app/Http/Controllers/PostsController.php
	modified:   resources/views/posts/show.blade.php

+++ app/Http/Controllers/PostsController.php
@@ -10,6 +10,17 @@ use DB;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        //$this->middleware('auth');
+        $this->middleware('auth',['except'=>['index','show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -87,6 +98,13 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //Check for crt usr
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -124,6 +142,12 @@ class PostsController extends Controller
     {
         //
         $post = Post::find($id);
+
+        //Check for crt usr
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized Page');
+        }
+        
         $post->delete();
         return redirect('/posts')->with('success','Article Deleted');
     }
